Started on title screen, adjusted fireball bbox width.
The fireball's bounding box has been made wider than the image to
compensate for the fact that fireballs can move at speeds of up to
32; moving faster than the width of the bounding box can cause
collision glitches. (In particular, this caused the fireball to hit
a wall behind the top of a slope.)

In the future, when I have a faster SGE implementation, I'll revert
this and instead limit delta timing to 30 FPS. But at the moment,
there's only the Pygame SGE, which doesn't handle 30 FPS well enough.