Fixed a problem causing iceblocks to be automatically kicked.
I recently added in some code to make Tux automatically kick a
second iceblock if already holding one, but I failed to check if
an iceblock being collided with was actually a different one. This
caused the iceblock to be auto-kicked if an obstruction prevented
it from moving with Tux long enough to bring it out of contact,
because a collision then happened.