Fixed a problem with tile deactivation.
Basically, it seems like when the player was instantly teleported
from one area to another, e.g. because of warp pipes, tiles were
deactivated first, and this resulted in a frame where enemies could
start falling through. Fixed this by not actually making a tile
intangible until it's been outside view for 3 frames.