Return to the map when you lose the level.
Turns out I had a defect I had to fix along the way: the
newer method of controlling map movement stopped working after you
left the map and came back. So I replaced that with a more
thorough version of the previous version; instead of just assuming
that empty spaces are "solved", they're considered to be "solved"
if they are connected, either directly or indirectly, to a "solved"
space that actually has a level.