A bunch of stuff mostly related to cutscenes.
Forgot to make a commit earlier. So, this removes win cutscenes,
since those can simply be made the end of the final level, and should
be. Adds a function which would be called when the credits
should roll; currently it just returns you to the title screen.
Added a file which is going to become the intro cutscene for reTux.
Added "steady icicle", which is going to be used for Raccot later:
a variant of icicles that doesn't automatically make it self fall and
instead has to be made to fall by an external force.