Adjusted the way searching for Tux dolls is handled.
Now, the loading bar refers to the action as "preloading levels",
and it indeed now preloads every level entirely, unless the process
is interrupted by the user. In case this happens, there are two
fallbacks for determining number of Tux dolls: one, it can be
manually indicated in the levelset file. Two, it will check levels
for Tux dolls any time they are loaded.

This is an attempt to continue to gain from the benefits of preloading
levels without annoying the users. Players who want to just get
started playing the game immediately can skip the preloading and
accept the resultant delays in loading new worldmaps. Players who
want the experience to be more smooth can wait for it to preload
(which, depending on the size of the game, may take as long as a minute).