Made objects left on moving platforms stay on them.
Part of this required removing code that caused walls to become
intangible if they were far outside of the view. Well, disabling it,
rather. This is the second time I've done this; it's just a terribly
buggy piece of code, and doesn't cause enough benefit to be worth
it. (Another problem it was causing was occasionally making objects
fall through the floor when re-entering rooms.)