Fixed a bug in the credits screen.
It was caused by the automatic caching behavior for levels: the
same room was being loaded, and on top of that, it was re-generating
credits because that code was in event_room_resume instead of
event_room_start. Worked around by checking for the presence of the
credits screen in either of the two caches and deleting it from them
if necessary.