Added flying spiky sprite.
The flying spiky is something I've been thinking could really be
useful for quite some time. I'll just need to pay for someone to
make a proper "iced" sprite for it, and maybe make the blade more
consistent with the rest of the sprite (by adding the same sort of
semi-outline).