Fixed a bug that caused multiple paths to be followed on the maps.
The result of this bug was that seemingly random paths around the
map were sometimes chosen. This is a very strange behavior, and I
don't understand why it failed in that particular way, but the root
cause was that the code permitted attempting to follow multiple
paths at once. This has been fixed by actually checking MapPlayer's
"moving" attribute.