Only skip the death sequence if the Escape key is pressed.
Making it any key causes this to inadvertantly happen just about
every time it happens, which is unfortunate aesthetically. Some
people might get the impression that there is no animation, that
it just makes a noise and returns to the map.

I did have a reason for implementing it as any key: SuperTux had
been annoying me by having me reach over to press the Escape key
each time I died, and the dying was happening a lot. But considering
the aesthetic cost, I don't think it's really worth it.