Changed rocks from SolidTop to Solid.
This generally has a positive effect. It makes it less likely for
them to take up the same spot, which can slow down the game and
reduces their effectiveness.

I was concerned that this could trap the player, but tests show that
all the player needs to do is pick up and drop (or throw) rocks in
succession, because this displaces them somewhat.