Simplified the dependencies listed in the readme.
I'm going to distribute the Pygame SGE with this from now on. It's
just a bit much to ask for people to download the Pygame SGE
themselves, especially on Windows. Now, I'm only telling users to
install what I just can't realistically distribute with the game
for everyone (i.e. the one library that isn't pure-Python, and
Python itself).