Fixed a problem with explosions, added a fire flower to level 26.
The problem was that the explosion was doing nothing to hittable
blocks if it didn't have a "detonator". Whoops! Fixed now by picking
the nearest player at the time of impact. If there are no players
at all, it passes "None" as the "other" for the block (which now
has the effect of the coins being lost to the void).