Improved the way rocks work.
They've now been split into two objects: the rock itself, which is
an InteractiveObject, and a child wall which is moved along with
it. This allows the wall to be intangible while the rock is being
carried without the side effect of being able to make the rock go
through other walls and slopes.

Incidentally, this also avoids slowing down the game as much.