Increased the gravity level a bit.
My primary motivation for this change was to make long jumps
shorter, since the distance of long jumps was previously ridiculously
long. However, this also has the effect of making Tux's jumps less
"floaty", and making the time Jumpies and bouncing snowballs are in
the air shorter, which makes them just slightly harder to deal with.

This made the game feel faster, so to counteract this, I slowed down
the target FPS just a bit (to 54 instead of 60). This makes some
timing things shorter (like explosions), but not not to an extent
that's particularly noticeable.