Added a check to make sure the player can normally move to a door's position.
This is to make absolutely sure that the player can exit the door
that was entered (unless the level is designed to prevent this
intentionally). It was previously technically possible, though highly
unlikely, for temporary blocks to fail to block entrance but end up
blocking the exit anyway. (Highly unlikely because in practice, this
requires very specialized design.)