Added intelligent inclusion of all obviously necessary files in a levelset.
This turned out to be really difficult, and I was debugging this over the
last two days. But it's done now. :)

The automatic inclusion does not include any files found in the base
installation of ReTux. It also does not examine timelines for possible
extra files, so those still need to be specified manually.