Made a small tweak which I think fixes something.
There's a minor bug where skipping through preloading would cause
even the current level to not be preloaded and result in a delay
where the top-left of the map was visible and the "loading level"
text was shown. I think this should fix it: I just moved the code
which accepts input for skipping preloading to after a level segment
is loaded, not before.