Reduced the crusher and icicle lax so that certain sections work.
Basically, there are a bunch of places where the player is supposed
to run underneath a crusher and pass it, but this depended on the
old bug which reduced the crusher lax arbitrarily. This small change
makes those sections work properly again (in fact, makes them work
more consistently than before) without ruining other cases.