Fixed a possibility of removing a keyboard control.
The code previously added to prevent this was incomplete; it could
still happen by assigning the only control to something else.

Since there are several different buttons assigned to "enter" for
menus, this should not realistically be a problem big enough to
warrant a new release just for this fix.