Fixed the crushers' response to having a rock placed on top of them.
I think it's actually kind of funny that the rocks stop them from
moving up, but they were also stopping them from moving up even after
they were removed, which could in some cases make them stuck on the
ground forever. I've remedied this by adding code that makes it
start moving up if it turns out that it's not on a ceiling.