Halved ice flower animation rate.
Bit of an experiment: this is designed to make it easier for
colorblind players to see a distinction between the fire flower and
the ice flower. A distinction can already be seen (the lightness of
the leaves; ice flower leaves are brighter than fire flower leaves),
but this might make it easier for those who are fully colorblind.