Removed time penalty.
This is something that someone brought up years ago: time penalty is
something that, while it doesn't really affect the result all that
much, has a major psychological impact that can be demoralizing,
since it feels like you're being punished. So I've changed it so that
time penalty can never happen. The mechanics are still there and
supported by the game, but they simply will never be reached
anymore because the time bonus is never allowed to go below zero.