Improved the final boss.
This started out as just wanting to make the cutscene better, but
I ended up overhauling the entire thing… and now it's two hours
later. 😅

The cutscene has been split apart into two smaller cutscene: the
first cutscene proceeds as the original cutscene did, but when
Raccot finishes moving, instead of making Tux walk to the right,
it returns control to the player. Then, a second cutscene is triggered
when the player reaches the middle of the arena: this cutscene makes
Raccot jump up and smash the two boxes, closes the boss door, and then
starts up the boss sequence. There's no longer an attempt to
synchronize things with the start of the music.

After I finished that, I decided to improve the gameplay of the arena
itself. I wanted to get rid of that mess of coins that was there.
So what I ended up doingg is having only a few coins, plus some in
the passageway again, and then some infinitely spawning coins coming
out of pipes which are now placed at the edges of the arena. Based on
my tests, it seems that these coins make the boss much, much easier.
Each pipe will spawn a total of 15 coins (5 in each position) before
stopping; one coin spawns every 90 frames (about 1.5 seconds). What
this does is allow the player to regularly get large supplies of coins
to refill their health, making the boss much, much more forgiving.