Ensure that death fadeouts is smooth nearly all of the time.
This is a workaround for an issue where a start track can still be
at the start portion when the death sequence fades it out. In that
case, fading  out near the end of the track would cause the music
to seemingly cut off abruptly. To avoid this, we shorten the
fade-out time if the remaining time on the currently playing music
is less than the default fade-out time.

This will have a side-effect where dying during the end of the loop
portion will cause it to fade out faster than expected, but it should
be rare for this to be problematic.