Made level 37 less strict, prevent bad music stops with looping part.
Accidentally merged these two into one commit. Oops. In any case,
this does two things: firstly, it makes it so that level 37's bombs
that are used for the Tux doll puzzle are spawned infinitely from a
pipe, ensuring you have infinite tries. Secondly, it adds a conditional
to ensure that the looping portion of music always fades out with the
full fade-out time (i.e. the shortening based on length is only used
for the start portions).