Adjusted the dark room's layout in Danger Doors so it's more fair.
Two major changes: firstly, I moved a platform so that the next one
after it isn't impossible to see with the lantern. Secondly, I added
a bit of a divet where Krush falls so that you don't get blindsided
by it before you can discover its presence. It's still appropriately
difficult to try to sneak through with just the fire flower.