Added support for using the left and right arrows in the menus.
The left and right arrows are now usable! For most menus this just
does the same as the enter key but for the options in the options
menu it allows cycling backwards (e.g. lowering the volume), like
in SuperTux.

also changed the use of < and > for selection markers to now use
« and » instead. Apparently these are used as quotation marks in
some locales, which makes them a perfect choice I think.