Added a capability to get rid of unwanted second controls, fixed old text.
This causes all existing controls to erase if you try to re-bind a
control that's already bound to an action, now allowing for an action
to only have one control bound to it.

Also fixes a bug where the old, outdated "toggle" text was still being
displayed for some reason (the updated text was just kind of ignored
as it turns out).