Adjusted level 21 to use its own bonus room instead of re-using bonus1.
The fact that both levels use the same room is a leftover from early
on when we had the idea to have a collection of bonus rooms that were
repeated similar to Super Mario Bros. It ended up not really being
a thing, and with level 21 being a hard level (bot designated and in
practice), it was just a bit jarring to see level 1's bonus room at the
end of it. This new bonus room is more interesting because it gives you
an iceblock to throw around and collect tons of coins from the bricks.