Reworked the infoboxes' text, updated translation files.
I made it so that all infoboxes in the ReTux levelset are either
"Traveler's Note" or "Minion's Note" boxes, in an attempt to make
some of them seem less unusual and almost "omniscient". Also updated
translation files, taking care to ensure that old translations map up
to the text that replaces what they originally translated.

Also reworded the dialog in the first castle, which I've felt was
rather clumsily worded for many years at this point.